+++ app/src/main/java/hexlet/code/repository/UrlCheckRepository.java
@@ -16,26 +16,26 @@ public class UrlCheckRepository extends BaseRepository {
             preparedStatement.setString(2, urlCheck.getTitle());
             preparedStatement.setString(3, urlCheck.getH1());
             preparedStatement.setString(4, urlCheck.getDescription());
-            preparedStatement.setInt(5, urlCheck.getUrlId());
+            preparedStatement.setLong(5, urlCheck.getUrlId());
             preparedStatement.executeUpdate();
             var generatedKeys = preparedStatement.getGeneratedKeys();
             if (generatedKeys.next()) {
-                urlCheck.setId(generatedKeys.getInt(1));
+                urlCheck.setId(generatedKeys.getLong(1));
             } else {
                 throw new SQLException("DB have not returned an id after saving an entity");
             }
         }
     }
 
-    public static List<UrlCheck> findByUrlId(Integer urlId) throws SQLException {
+    public static List<UrlCheck> findByUrlId(Long urlId) throws SQLException {
         var sql = "SELECT * FROM url_checks WHERE url_id = ?";
         try (var conn = dataSource.getConnection();
              var stmt = conn.prepareStatement(sql)) {
-            stmt.setInt(1, urlId);
+            stmt.setLong(1, urlId);
             var resultSet = stmt.executeQuery();
             var result = new ArrayList<UrlCheck>();
             while (resultSet.next()) {
-                var id = resultSet.getInt("id");
+                var id = resultSet.getLong("id");
                 var statusCode = resultSet.getInt("status_code");
                 var title = resultSet.getString("title");
                 var h1 = resultSet.getString("h1");
@@ -51,7 +51,7 @@ public class UrlCheckRepository extends BaseRepository {
         }
     }
 
-    public static List<UrlCheck> findLatestForUrls(List<Integer> urlIds) throws SQLException {
+    public static List<UrlCheck> findLatestForUrls(List<Long> urlIds) throws SQLException {
         var sql = "SELECT id, status_code, title, h1, description,"
                 + "max(created_at) as created_at, url_id FROM url_checks WHERE url_id = ANY (?) GROUP BY id, url_id";
         try (var conn = dataSource.getConnection();
@@ -60,13 +60,13 @@ public class UrlCheckRepository extends BaseRepository {
             var resultSet = stmt.executeQuery();
             var result = new ArrayList<UrlCheck>();
             while (resultSet.next()) {
-                var id = resultSet.getInt("id");
+                var id = resultSet.getLong("id");
                 var statusCode = resultSet.getInt("status_code");
                 var title = resultSet.getString("title");
                 var h1 = resultSet.getString("h1");
                 var description = resultSet.getString("description");
                 var createdAt = resultSet.getTimestamp("created_at");
-                var urlId = resultSet.getInt("url_id");
+                var urlId = resultSet.getLong("url_id");
                 var urlCheck = new UrlCheck(statusCode, title, h1, description);
                 urlCheck.setId(id);
                 urlCheck.setCreatedAt(createdAt);
